serverless/cdk-twilio-sms-discord/lambda/requirements.txt
twilio>=7.8.2
PyNaCl>=1.5.0
https://github.com/Rapptz/discord.py/archive/refs/heads/master.zip
